reorganize; add complete partial da code
