+++ dalib/vision/datasets/__init__.py
@@ -4,6 +4,8 @@ from .officehome import OfficeHome
 from .visda2017 import VisDA2017
 from .officecaltech import OfficeCaltech
 from .domainnet import DomainNet
-from .partialda import PartialOffice31
+from .imagenet_caltech import ImageNetCaltech
+from .caltech_imagenet import CaltechImageNet
 
-__all__ = ['ImageList', 'Office31', 'OfficeHome', "VisDA2017", "OfficeCaltech", "DomainNet", "PartialOffice31"]
+
+__all__ = ['ImageList', 'Office31', 'OfficeHome', "VisDA2017", "OfficeCaltech", "DomainNet", 'ImageNetCaltech', 'CaltechImageNet']
