--- dalib/vision/datasets/imagenet.py
@@ -1,1 +1,0 @@
-import torchvision.datasets
