+++ examples/unsupervised-da/jan.py
@@ -6,7 +6,6 @@ import argparse
 import copy
 
 import torch
-import torch.nn as nn
 import torch.nn.parallel
 import torch.backends.cudnn as cudnn
 from torch.optim import SGD
@@ -17,14 +16,15 @@ import torchvision.transforms as transforms
 import torch.nn.functional as F
 
 sys.path.append('.')
-from dalib.modules.domain_discriminator import DomainDiscriminator
-from dalib.adaptation.cdan import ConditionalDomainAdversarialLoss, ImageClassifier
+from dalib.adaptation.jan import JointMultipleKernelMaximumMeanDiscrepancy, ImageClassifier, Theta
+from dalib.modules.kernels import GaussianKernel
 import dalib.vision.datasets as datasets
 import dalib.vision.models as models
-from tools.utils import AverageMeter, ProgressMeter, accuracy, ForeverDataIterator
-from tools.transforms import ResizeImage
-from tools.lr_scheduler import StepwiseLR
-
+from dalib.vision.transforms import ResizeImage
+from dalib.utils.data import ForeverDataIterator
+from dalib.utils.metric import accuracy
+from dalib.utils.avgmeter import AverageMeter, ProgressMeter
+from dalib.optim.lr_scheduler import StepwiseLR
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
@@ -68,7 +68,7 @@ def main(args: argparse.Namespace):
     val_dataset = dataset(root=args.root, task=args.target, download=True, transform=val_tranform)
     val_loader = DataLoader(val_dataset, batch_size=args.batch_size, shuffle=False, num_workers=args.workers)
     if args.data == 'DomainNet':
-        test_dataset = dataset(root=args.root, task=args.target, evaluate=True, download=True, transform=val_tranform)
+        test_dataset = dataset(root=args.root, task=args.target, split='test', download=True, transform=val_tranform)
         test_loader = DataLoader(test_dataset, batch_size=args.batch_size, shuffle=False, num_workers=args.workers)
     else:
         test_loader = val_loader
@@ -81,27 +81,33 @@ def main(args: argparse.Namespace):
     backbone = models.__dict__[args.arch](pretrained=True)
     num_classes = train_source_dataset.num_classes
     classifier = ImageClassifier(backbone, num_classes).to(device)
-    classifier_feature_dim = classifier.features_dim
-
-    domain_discri = DomainDiscriminator(classifier_feature_dim * num_classes, hidden_size=1024).to(device)
-
-    all_parameters = classifier.get_parameters() + domain_discri.get_parameters()
-
-    # define optimizer and lr scheduler
-    optimizer = SGD(all_parameters, args.lr, momentum=args.momentum, weight_decay=args.weight_decay, nesterov=True)
-    lr_sheduler = StepwiseLR(optimizer, init_lr=args.lr, gamma=0.001, decay_rate=0.75)
 
     # define loss function
-    domain_adv = ConditionalDomainAdversarialLoss(
-        domain_discri, entropy_conditioning=args.entropy,
-        num_classes=num_classes, features_dim=classifier_feature_dim, randomized=False
+    if args.adversarial:
+        thetas = [Theta(dim).to(device) for dim in (classifier.features_dim, num_classes)]
+    else:
+        thetas = None
+    jmmd_loss = JointMultipleKernelMaximumMeanDiscrepancy(
+        kernels=(
+            [GaussianKernel(alpha=2 ** k) for k in range(-3, 2)],
+            (GaussianKernel(sigma=0.92, track_running_stats=False),)
+        ),
+        linear=args.linear, thetas=thetas
     ).to(device)
+
+    parameters = classifier.get_parameters()
+    if thetas is not None:
+        parameters += [{"params": theta.parameters(), 'lr_mult': 0.1} for theta in thetas]
+
+    # define optimizer
+    optimizer = SGD(parameters, args.lr, momentum=args.momentum, weight_decay=args.wd, nesterov=True)
+    lr_sheduler = StepwiseLR(optimizer, init_lr=args.lr, gamma=0.0003, decay_rate=0.75)
 
     # start training
     best_acc1 = 0.
     for epoch in range(args.epochs):
         # train for one epoch
-        train(train_source_iter, train_target_iter, classifier, domain_adv, optimizer,
+        train(train_source_iter, train_target_iter, classifier, jmmd_loss, optimizer,
               lr_sheduler, epoch, args)
 
         # evaluate on validation set
@@ -121,22 +127,23 @@ def main(args: argparse.Namespace):
 
 
 def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverDataIterator, model: ImageClassifier,
-          domain_adv: ConditionalDomainAdversarialLoss, optimizer: SGD,
+          jmmd_loss: JointMultipleKernelMaximumMeanDiscrepancy, optimizer: SGD,
           lr_sheduler: StepwiseLR, epoch: int, args: argparse.Namespace):
-    batch_time = AverageMeter('Time', ':3.1f')
+    batch_time = AverageMeter('Time', ':4.2f')
     data_time = AverageMeter('Data', ':3.1f')
     losses = AverageMeter('Loss', ':3.2f')
-    trans_losses = AverageMeter('Trans Loss', ':3.2f')
+    trans_losses = AverageMeter('Trans Loss', ':5.4f')
     cls_accs = AverageMeter('Cls Acc', ':3.1f')
-    domain_accs = AverageMeter('Domain Acc', ':3.1f')
+    tgt_accs = AverageMeter('Tgt Acc', ':3.1f')
+
     progress = ProgressMeter(
         args.iters_per_epoch,
-        [batch_time, data_time, losses, trans_losses, cls_accs, domain_accs],
+        [batch_time, data_time, losses, trans_losses, cls_accs, tgt_accs],
         prefix="Epoch: [{}]".format(epoch))
 
     # switch to train mode
     model.train()
-    domain_adv.train()
+    jmmd_loss.train()
 
     end = time.time()
     for i in range(args.iters_per_epoch):
@@ -146,11 +153,12 @@ def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverData
         data_time.update(time.time() - end)
 
         x_s, labels_s = next(train_source_iter)
-        x_t, _ = next(train_target_iter)
+        x_t, labels_t = next(train_target_iter)
 
         x_s = x_s.to(device)
         x_t = x_t.to(device)
         labels_s = labels_s.to(device)
+        labels_t = labels_t.to(device)
 
         # compute output
         x = torch.cat((x_s, x_t), dim=0)
@@ -159,15 +167,18 @@ def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverData
         f_s, f_t = f.chunk(2, dim=0)
 
         cls_loss = F.cross_entropy(y_s, labels_s)
-        transfer_loss = domain_adv(y_s, f_s, y_t, f_t)
-        domain_acc = domain_adv.domain_discriminator_accuracy
+        transfer_loss = jmmd_loss(
+            (f_s, F.softmax(y_s, dim=1)),
+            (f_t, F.softmax(y_t, dim=1))
+        )
         loss = cls_loss + transfer_loss * args.trade_off
 
         cls_acc = accuracy(y_s, labels_s)[0]
+        tgt_acc = accuracy(y_t, labels_t)[0]
 
         losses.update(loss.item(), x_s.size(0))
         cls_accs.update(cls_acc.item(), x_s.size(0))
-        domain_accs.update(domain_acc.item(), x_s.size(0))
+        tgt_accs.update(tgt_acc.item(), x_t.size(0))
         trans_losses.update(transfer_loss.item(), x_s.size(0))
 
         # compute gradient and do SGD step
@@ -256,23 +267,26 @@ if __name__ == '__main__':
     parser.add_argument('-b', '--batch-size', default=32, type=int,
                         metavar='N',
                         help='mini-batch size (default: 32)')
-    parser.add_argument('--lr', '--learning-rate', default=0.01, type=float,
+    parser.add_argument('--lr', '--learning-rate',  default=0.003, type=float,
                         metavar='LR', help='initial learning rate', dest='lr')
     parser.add_argument('--momentum', default=0.9, type=float, metavar='M',
                         help='momentum')
-    parser.add_argument('--wd', '--weight-decay', default=1e-3, type=float,
-                        metavar='W', help='weight decay (default: 1e-3)',
-                        dest='weight_decay')
+    parser.add_argument('--wd', '--weight-decay', default=0.0005, type=float,
+                        metavar='W', help='weight decay (default: 5e-4)')
     parser.add_argument('-p', '--print-freq', default=100, type=int,
                         metavar='N', help='print frequency (default: 100)')
-    parser.add_argument('--seed', default=0, type=int,
+    parser.add_argument('--seed', default=None, type=int,
                         help='seed for initializing training. ')
     parser.add_argument('--trade-off', default=1., type=float,
                         help='the trade-off hyper-parameter for transfer loss')
-    parser.add_argument('-i', '--iters-per-epoch', default=1000, type=int,
+    parser.add_argument('-i', '--iters-per-epoch', default=500, type=int,
                         help='Number of iterations per epoch')
-    parser.add_argument('--entropy', default=False, action='store_true', help='use entropy conditioning')
+    parser.add_argument('--linear', default=False, action='store_true',
+                        help='whether use the linear version')
+    parser.add_argument('--adversarial', default=False, action='store_true',
+                        help='whether use adversarial theta')
 
     args = parser.parse_args()
     print(args)
     main(args)
+
