+++ examples/unsupervised-da/mcd.py
@@ -4,10 +4,13 @@ import warnings
 import sys
 import argparse
 import copy
+from typing import Tuple
 
 import torch
+import torch.nn as nn
 import torch.nn.parallel
 import torch.backends.cudnn as cudnn
+import torch.optim
 from torch.optim import SGD
 import torch.utils.data
 from torch.utils.data import DataLoader
@@ -16,13 +19,14 @@ import torchvision.transforms as transforms
 import torch.nn.functional as F
 
 sys.path.append('.')
-from dalib.adaptation.jan import JointMultipleKernelMaximumMeanDiscrepancy, ImageClassifier, Theta
-from dalib.modules.kernels import GaussianKernel
+from dalib.adaptation.mcd import ImageClassifierHead, entropy, classifier_discrepancy
 import dalib.vision.datasets as datasets
 import dalib.vision.models as models
-from tools.utils import AverageMeter, ProgressMeter, accuracy, ForeverDataIterator
-from tools.transforms import ResizeImage
-from tools.lr_scheduler import StepwiseLR
+from dalib.vision.transforms import ResizeImage
+from dalib.utils.data import ForeverDataIterator
+from dalib.utils.metric import accuracy
+from dalib.utils.avgmeter import AverageMeter, ProgressMeter
+from dalib.optim.lr_scheduler import StepwiseLR
 
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -43,13 +47,22 @@ def main(args: argparse.Namespace):
 
     # Data loading code
     normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
-    train_transform = transforms.Compose([
-        ResizeImage(256),
-        transforms.RandomResizedCrop(224),
-        transforms.RandomHorizontalFlip(),
-        transforms.ToTensor(),
-        normalize
-    ])
+    if args.center_crop:
+        train_transform = transforms.Compose([
+            ResizeImage(256),
+            transforms.CenterCrop(224),
+            transforms.RandomHorizontalFlip(),
+            transforms.ToTensor(),
+            normalize
+        ])
+    else:
+        train_transform = transforms.Compose([
+            ResizeImage(256),
+            transforms.RandomResizedCrop(224),
+            transforms.RandomHorizontalFlip(),
+            transforms.ToTensor(),
+            normalize
+        ])
     val_tranform = transforms.Compose([
         ResizeImage(256),
         transforms.CenterCrop(224),
@@ -67,7 +80,7 @@ def main(args: argparse.Namespace):
     val_dataset = dataset(root=args.root, task=args.target, download=True, transform=val_tranform)
     val_loader = DataLoader(val_dataset, batch_size=args.batch_size, shuffle=False, num_workers=args.workers)
     if args.data == 'DomainNet':
-        test_dataset = dataset(root=args.root, task=args.target, evaluate=True, download=True, transform=val_tranform)
+        test_dataset = dataset(root=args.root, task=args.target, split='test', download=True, transform=val_tranform)
         test_loader = DataLoader(test_dataset, batch_size=args.batch_size, shuffle=False, num_workers=args.workers)
     else:
         test_loader = val_loader
@@ -77,61 +90,50 @@ def main(args: argparse.Namespace):
 
     # create model
     print("=> using pre-trained model '{}'".format(args.arch))
-    backbone = models.__dict__[args.arch](pretrained=True)
+    G = models.__dict__[args.arch](pretrained=True).to(device)  # feature extractor
     num_classes = train_source_dataset.num_classes
-    classifier = ImageClassifier(backbone, num_classes).to(device)
-
-    # define loss function
-    if args.adversarial:
-        thetas = [Theta(dim).to(device) for dim in (classifier.features_dim, num_classes)]
-    else:
-        thetas = None
-    jmmd_loss = JointMultipleKernelMaximumMeanDiscrepancy(
-        kernels=(
-            [GaussianKernel(alpha=2 ** k) for k in range(-3, 2)],
-            (GaussianKernel(sigma=0.92, track_running_stats=False),)
-        ),
-        linear=args.linear, thetas=thetas
-    ).to(device)
-
-    parameters = classifier.get_parameters()
-    if thetas is not None:
-        parameters += [{"params": theta.parameters(), 'lr_mult': 0.1} for theta in thetas]
+    # two image classifier heads
+    F1 = ImageClassifierHead(G.out_features, num_classes, args.bottleneck_dim).to(device)
+    F2 = ImageClassifierHead(G.out_features, num_classes, args.bottleneck_dim).to(device)
 
     # define optimizer
-    optimizer = SGD(parameters, args.lr, momentum=args.momentum, weight_decay=args.wd, nesterov=True)
-    lr_sheduler = StepwiseLR(optimizer, init_lr=args.lr, gamma=0.0003, decay_rate=0.75)
+    # the learning rate is fixed according to origin paper
+    optimizer_g = SGD(G.parameters(), lr=args.lr, weight_decay=0.0005)
+    optimizer_f = SGD(F1.get_parameters()+F2.get_parameters(), momentum=0.9, lr=args.lr, weight_decay=0.0005)
 
     # start training
     best_acc1 = 0.
+    best_results = None
     for epoch in range(args.epochs):
         # train for one epoch
-        train(train_source_iter, train_target_iter, classifier, jmmd_loss, optimizer,
-              lr_sheduler, epoch, args)
+        train(train_source_iter, train_target_iter, G, F1, F2, optimizer_g, optimizer_f, epoch, args)
 
         # evaluate on validation set
-        acc1 = validate(val_loader, classifier, args)
+        results = validate(val_loader, G, F1, F2, args)
 
         # remember best acc@1 and save checkpoint
-        if acc1 > best_acc1:
-            best_model = copy.deepcopy(classifier.state_dict())
-        best_acc1 = max(acc1, best_acc1)
+        if max(results) > best_acc1:
+            best_G, best_F1, best_F2 = copy.deepcopy(G.state_dict()), copy.deepcopy(F1.state_dict()), copy.deepcopy(F2.state_dict())
+            best_acc1 = max(results)
+            best_results = results
 
-    print("best_acc1 = {:3.1f}".format(best_acc1))
+    print("best_acc1 = {:3.1f}, results = {}".format(best_acc1, best_results))
 
     # evaluate on test set
-    classifier.load_state_dict(best_model)
-    acc1 = validate(test_loader, classifier, args)
-    print("test_acc1 = {:3.1f}".format(acc1))
+    G.load_state_dict(best_G)
+    F1.load_state_dict(best_F1)
+    F2.load_state_dict(best_F2)
+    results = validate(test_loader, G, F1, F2, args)
+    print("test_acc1 = {:3.1f}".format(max(results)))
 
 
-def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverDataIterator, model: ImageClassifier,
-          jmmd_loss: JointMultipleKernelMaximumMeanDiscrepancy, optimizer: SGD,
-          lr_sheduler: StepwiseLR, epoch: int, args: argparse.Namespace):
-    batch_time = AverageMeter('Time', ':4.2f')
+def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverDataIterator,
+          G: nn.Module, F1: ImageClassifierHead, F2: ImageClassifierHead,
+          optimizer_g: SGD, optimizer_f: SGD, epoch: int, args: argparse.Namespace):
+    batch_time = AverageMeter('Time', ':3.1f')
     data_time = AverageMeter('Data', ':3.1f')
     losses = AverageMeter('Loss', ':3.2f')
-    trans_losses = AverageMeter('Trans Loss', ':5.4f')
+    trans_losses = AverageMeter('Trans Loss', ':3.2f')
     cls_accs = AverageMeter('Cls Acc', ':3.1f')
     tgt_accs = AverageMeter('Tgt Acc', ':3.1f')
 
@@ -141,13 +143,12 @@ def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverData
         prefix="Epoch: [{}]".format(epoch))
 
     # switch to train mode
-    model.train()
-    jmmd_loss.train()
+    G.train()
+    F1.train()
+    F2.train()
 
     end = time.time()
     for i in range(args.iters_per_epoch):
-        lr_sheduler.step()
-
         # measure data loading time
         data_time.update(time.time() - end)
 
@@ -158,32 +159,61 @@ def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverData
         x_t = x_t.to(device)
         labels_s = labels_s.to(device)
         labels_t = labels_t.to(device)
-
-        # compute output
         x = torch.cat((x_s, x_t), dim=0)
-        y, f = model(x)
-        y_s, y_t = y.chunk(2, dim=0)
-        f_s, f_t = f.chunk(2, dim=0)
+        assert x.requires_grad is False
 
-        cls_loss = F.cross_entropy(y_s, labels_s)
-        transfer_loss = jmmd_loss(
-            (f_s, F.softmax(y_s, dim=1)),
-            (f_t, F.softmax(y_t, dim=1))
-        )
-        loss = cls_loss + transfer_loss * args.trade_off
+        # Step A train all networks to minimize loss on source domain
+        optimizer_g.zero_grad()
+        optimizer_f.zero_grad()
 
-        cls_acc = accuracy(y_s, labels_s)[0]
-        tgt_acc = accuracy(y_t, labels_t)[0]
+        g = G(x)
+        y_1 = F1(g)
+        y_2 = F2(g)
+        y1_s, y1_t = y_1.chunk(2, dim=0)
+        y2_s, y2_t = y_2.chunk(2, dim=0)
+
+        y1_t, y2_t = F.softmax(y1_t, dim=1), F.softmax(y2_t, dim=1)
+        loss = F.cross_entropy(y1_s, labels_s) + F.cross_entropy(y2_s, labels_s) + \
+               0.01 * (entropy(y1_t) + entropy(y2_t))
+        loss.backward()
+        optimizer_g.step()
+        optimizer_f.step()
+
+        # Step B train classifier to maximize discrepancy
+        optimizer_g.zero_grad()
+        optimizer_f.zero_grad()
+
+        g = G(x)
+        y_1 = F1(g)
+        y_2 = F2(g)
+        y1_s, y1_t = y_1.chunk(2, dim=0)
+        y2_s, y2_t = y_2.chunk(2, dim=0)
+        y1_t, y2_t = F.softmax(y1_t, dim=1), F.softmax(y2_t, dim=1)
+        loss = F.cross_entropy(y1_s, labels_s) + F.cross_entropy(y2_s, labels_s) + \
+               0.01 * (entropy(y1_t) + entropy(y2_t)) - classifier_discrepancy(y1_t, y2_t) * args.trade_off
+        loss.backward()
+        optimizer_f.step()
+
+        # Step C train genrator to minimize discrepancy
+        for k in range(args.num_k):
+            optimizer_g.zero_grad()
+            g = G(x)
+            y_1 = F1(g)
+            y_2 = F2(g)
+            y1_s, y1_t = y_1.chunk(2, dim=0)
+            y2_s, y2_t = y_2.chunk(2, dim=0)
+            y1_t, y2_t = F.softmax(y1_t, dim=1), F.softmax(y2_t, dim=1)
+            mcd_loss = classifier_discrepancy(y1_t, y2_t) * args.trade_off
+            mcd_loss.backward()
+            optimizer_g.step()
+
+        cls_acc = accuracy(y1_s, labels_s)[0]
+        tgt_acc = accuracy(y1_t, labels_t)[0]
 
         losses.update(loss.item(), x_s.size(0))
         cls_accs.update(cls_acc.item(), x_s.size(0))
         tgt_accs.update(tgt_acc.item(), x_t.size(0))
-        trans_losses.update(transfer_loss.item(), x_s.size(0))
-
-        # compute gradient and do SGD step
-        optimizer.zero_grad()
-        loss.backward()
-        optimizer.step()
+        trans_losses.update(mcd_loss.item(), x_s.size(0))
 
         # measure elapsed time
         batch_time.update(time.time() - end)
@@ -193,18 +223,20 @@ def train(train_source_iter: ForeverDataIterator, train_target_iter: ForeverData
             progress.display(i)
 
 
-def validate(val_loader: DataLoader, model: ImageClassifier, args: argparse.Namespace) -> float:
+def validate(val_loader: DataLoader, G: nn.Module, F1: ImageClassifierHead,
+             F2: ImageClassifierHead, args: argparse.Namespace) -> Tuple[float, float]:
     batch_time = AverageMeter('Time', ':6.3f')
-    losses = AverageMeter('Loss', ':.4e')
-    top1 = AverageMeter('Acc@1', ':6.2f')
-    top5 = AverageMeter('Acc@5', ':6.2f')
+    top1_1 = AverageMeter('Acc_1', ':6.2f')
+    top1_2 = AverageMeter('Acc_2', ':6.2f')
     progress = ProgressMeter(
         len(val_loader),
-        [batch_time, losses, top1, top5],
+        [batch_time, top1_1, top1_2],
         prefix='Test: ')
 
     # switch to evaluate mode
-    model.eval()
+    G.eval()
+    F1.eval()
+    F2.eval()
 
     with torch.no_grad():
         end = time.time()
@@ -213,14 +245,14 @@ def validate(val_loader: DataLoader, model: ImageClassifier, args: argparse.Name
             target = target.to(device)
 
             # compute output
-            output, _ = model(images)
-            loss = F.cross_entropy(output, target)
+            g = G(images)
+            y1, y2 = F1(g), F2(g)
 
             # measure accuracy and record loss
-            acc1, acc5 = accuracy(output, target, topk=(1, 5))
-            losses.update(loss.item(), images.size(0))
-            top1.update(acc1[0], images.size(0))
-            top5.update(acc5[0], images.size(0))
+            acc1, = accuracy(y1, target)
+            acc2, = accuracy(y2, target)
+            top1_1.update(acc1[0], images.size(0))
+            top1_2.update(acc2[0], images.size(0))
 
             # measure elapsed time
             batch_time.update(time.time() - end)
@@ -229,10 +261,10 @@ def validate(val_loader: DataLoader, model: ImageClassifier, args: argparse.Name
             if i % args.print_freq == 0:
                 progress.display(i)
 
-        print(' * Acc@1 {top1.avg:.3f} Acc@5 {top5.avg:.3f}'
-              .format(top1=top1, top5=top5))
+        print(' * Acc1 {top1_1.avg:.3f} Acc2 {top1_2.avg:.3f}'
+              .format(top1_1=top1_1, top1_2=top1_2))
 
-    return top1.avg
+    return top1_1.avg, top1_2.avg
 
 
 if __name__ == '__main__':
@@ -266,25 +298,20 @@ if __name__ == '__main__':
     parser.add_argument('-b', '--batch-size', default=32, type=int,
                         metavar='N',
                         help='mini-batch size (default: 32)')
-    parser.add_argument('--lr', '--learning-rate',  default=0.003, type=float,
+    parser.add_argument('--lr', '--learning-rate', default=0.001, type=float,
                         metavar='LR', help='initial learning rate', dest='lr')
-    parser.add_argument('--momentum', default=0.9, type=float, metavar='M',
-                        help='momentum')
-    parser.add_argument('--wd', '--weight-decay', default=0.0005, type=float,
-                        metavar='W', help='weight decay (default: 5e-4)')
+    parser.add_argument('--num-k', type=int, default=4, metavar='K',
+                        help='how many steps to repeat the generator update')
     parser.add_argument('-p', '--print-freq', default=100, type=int,
-                        metavar='N', help='print frequency (default: 100)')
-    parser.add_argument('--seed', default=0, type=int,
+                        metavar='N', help='print frequency (default: 10)')
+    parser.add_argument('--seed', default=None, type=int,
                         help='seed for initializing training. ')
+    parser.add_argument('-i', '--iters-per-epoch', default=1000, type=int,
+                        help='Number of iterations per epoch')
+    parser.add_argument('--bottleneck-dim', default=1024, type=int)
+    parser.add_argument('--center-crop', default=False, action='store_true')
     parser.add_argument('--trade-off', default=1., type=float,
                         help='the trade-off hyper-parameter for transfer loss')
-    parser.add_argument('-i', '--iters-per-epoch', default=500, type=int,
-                        help='Number of iterations per epoch')
-    parser.add_argument('--linear', default=False, action='store_true',
-                        help='whether use the linear version')
-    parser.add_argument('--adversarial', default=False, action='store_true',
-                        help='whether use adversarial theta')
-
     args = parser.parse_args()
     print(args)
     main(args)
