--- tools/utils.py
@@ -1,98 +1,0 @@
-import sys
-import torch
-import os
-import shutil
-from torch.utils.data.dataloader import DataLoader
-sys.path.append('.')
-
-
-class AverageMeter(object):
-    """Computes and stores the average and current value"""
-    def __init__(self, name, fmt=':f'):
-        self.name = name
-        self.fmt = fmt
-        self.reset()
-
-    def reset(self):
-        self.val = 0
-        self.avg = 0
-        self.sum = 0
-        self.count = 0
-
-    def update(self, val, n=1):
-        self.val = val
-        self.sum += val * n
-        self.count += n
-        self.avg = self.sum / self.count
-
-    def __str__(self):
-        fmtstr = '{name} {val' + self.fmt + '} ({avg' + self.fmt + '})'
-        return fmtstr.format(**self.__dict__)
-
-
-class ProgressMeter(object):
-    def __init__(self, num_batches, meters, prefix=""):
-        self.batch_fmtstr = self._get_batch_fmtstr(num_batches)
-        self.meters = meters
-        self.prefix = prefix
-
-    def display(self, batch):
-        entries = [self.prefix + self.batch_fmtstr.format(batch)]
-        entries += [str(meter) for meter in self.meters]
-        print('\t'.join(entries))
-
-    def _get_batch_fmtstr(self, num_batches):
-        num_digits = len(str(num_batches // 1))
-        fmt = '{:' + str(num_digits) + 'd}'
-        return '[' + fmt + '/' + fmt.format(num_batches) + ']'
-
-
-def accuracy(output, target, topk=(1,)):
-    """Computes the accuracy over the k top predictions for the specified values of k"""
-    with torch.no_grad():
-        maxk = max(topk)
-        batch_size = target.size(0)
-
-        _, pred = output.topk(maxk, 1, True, True)
-        pred = pred.t()
-        correct = pred.eq(target.view(1, -1).expand_as(pred))
-
-        res = []
-        for k in topk:
-            correct_k = correct[:k].view(-1).float().sum(0, keepdim=True)
-            res.append(correct_k.mul_(100.0 / batch_size))
-        return res
-
-
-def create_exp_dir(path, scripts_to_save=None):
-    os.makedirs(path, exist_ok=True)
-
-    print('Experiment dir : {}'.format(path))
-
-    if scripts_to_save is not None:
-        script_path = os.path.join(path, 'scripts')
-        if os.path.exists(script_path):
-            shutil.rmtree(script_path)
-        os.mkdir(script_path)
-        for script in scripts_to_save:
-            dst_file = os.path.join(path, 'scripts', os.path.basename(script))
-            print(dst_file)
-            shutil.copytree(script, dst_file)
-
-
-class ForeverDataIterator:
-    """A data iterator that will never stop producing data"""
-    def __init__(self, data_loader: DataLoader):
-        self.data_loader = data_loader
-        self.iter = iter(self.data_loader)
-
-    def __next__(self):
-        try:
-            data = next(self.iter)
-        except StopIteration:
-            self.iter = iter(self.data_loader)
-            data = next(self.iter)
-        return data
-
-    def __len__(self):
-        return len(self.data_loader)
